dt clamping to avoid lag spikes, drop vsync support, round moving offset in entities, fix tile tearing issue

+++ include/Tiles/Tile.hxx
@@ -7,7 +7,7 @@ class Tile : public TileBase
   private:
   public:
     Tile(const std::string name, const std::string &id, sf::Texture &texture, const sf::IntRect &texture_rect,
-         const unsigned int &grid_size, const sf::Vector2u &grid_position, const float &scale,
+         const unsigned int &grid_size, const sf::Vector2u &grid_position, const unsigned int &scale,
          const sf::Color &color = sf::Color::White);
 
     ~Tile();
